Fix validation, remove unused files
